requirements.txt
aiosqlite==0.17.0
alembic==1.7.7
anyio==3.7.1
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.78.0
greenlet==3.0.0
h11==0.14.0
httptools==0.6.1
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
pydantic==1.10.13
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==1.4.36
starlette==0.19.1
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.17.6
uvloop==0.18.0
watchgod==0.8.2
websockets==11.0.3
